+++ src/UserInterface/Components/MainPanel/DateTime/SubComponents/NeonCircle.h
@@ -6,14 +6,14 @@ class NeonCircle : public MainComponent {
 
 	public:
 
-		void render(uint8_t value);
+		void render(int32_t value);
 		NeonCircle(uint8_t x, uint8_t y, uint8_t maxValue, long color);
 
 	protected:
 
-		uint8_t x, y;
+		int32_t x, y;
 		long color;
-		uint8_t prevValue;
-		uint8_t maxValue;
+		int32_t prevValue;
+		int32_t maxValue;
 
 };
