+++ src/Utils/Geometry.cpp
@@ -9,8 +9,8 @@ void Geometry::calculatePointOnCircle(
 	uint8_t y,
 	int16_t angle,
 	uint8_t radius,
-	uint8_t &x1,
-	uint8_t &y1
+	int32_t &x1,
+	int32_t &y1
 ) {
 	angle = angle - 90;
 	float a = angle * M_PI / 180;
